[REFACTOR] 파일 구조 수정, css 파일 분리
