+++ vestie/src/component/completeproceedsurvey/style.css
@@ -1,0 +1,26 @@
+.complete_survey_box {
+
+    margin-top: 30vh;
+    display: flex;
+    flex-direction: column;
+}
+
+.complete_survey_content {
+    font-size: 25px;
+    font-weight: bold;
+    margin: 0 auto;
+}
+.icon_box {
+    margin: 0 auto;
+}
+.complete_survey_gave_point_box {
+    margin: 0 auto;
+    font-size: 20px;
+    color: #929292;
+}
+.complete_survey_gave_point {
+    font-size: 30px;
+    font-weight: bold;
+    color: #4184F2;
+    margin-right: 3px;
+}
