+++ vestie/src/component/surveyitem/style.css
@@ -1,7 +1,3 @@
-body {
-    background-color: #F6F6F6;
-
-}
 .survey_item_box {
     background-color: white;
     border-radius: 10px;
