+++ vestie/src/page/main/main.css
@@ -1,0 +1,4 @@
+body {
+    background-color: #F6F6F6;
+
+}
