+++ vestie/src/page/proceedsurvey/proceedSurvey.css
@@ -1,0 +1,6 @@
+body {
+    background-color: #F6F6F6;
+    align-items: center;
+    justify-content: center;
+
+}
